pyproject.toml
[project]
name = "test-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.5.0",
    "python-dotenv>=1.0.1",
]
